[Issue #9316] Move workflow tests to correct folder
