+++ api/tests/src/api/workflows/test_workflow_audit_routes.py
@@ -11,7 +11,7 @@ from tests.src.db.models.factories import (
     WorkflowEventHistoryFactory,
     WorkflowFactory,
 )
-from tests.workflow.state_machine.test_state_machines import BasicState
+from tests.src.workflow.state_machine.test_state_machines import BasicState
 
 ####################################
 # Fixtures
