+++ api/tests/src/workflow/handler/test_event_handler.py
@@ -14,8 +14,8 @@ from src.workflow.workflow_errors import (
     WorkflowDoesNotExistError,
 )
 from tests.src.db.models.factories import OpportunityFactory, UserFactory, WorkflowFactory
-from tests.workflow.state_machine.test_state_machines import BasicState, BasicTestStateMachine
-from tests.workflow.workflow_test_util import (
+from tests.src.workflow.state_machine.test_state_machines import BasicState, BasicTestStateMachine
+from tests.src.workflow.workflow_test_util import (
     build_process_workflow_event,
     build_start_workflow_event,
 )
