+++ api/tests/src/workflow/listener/test_workflow_audit_listener.py
@@ -4,8 +4,8 @@ from src.constants.lookup_constants import ApprovalResponseType, WorkflowType
 from src.db.models.workflow_models import WorkflowAudit
 from src.workflow.handler.event_handler import EventHandler
 from tests.src.db.models.factories import OpportunityFactory, UserFactory, WorkflowFactory
-from tests.workflow.state_machine.test_state_machines import BasicState
-from tests.workflow.workflow_test_util import (
+from tests.src.workflow.state_machine.test_state_machines import BasicState
+from tests.src.workflow.workflow_test_util import (
     build_process_workflow_event,
     build_start_workflow_event,
 )
