+++ api/tests/src/workflow/manager/test_workflow_manager.py
@@ -21,8 +21,8 @@ from src.workflow.manager.workflow_manager import (
     handle_event,
 )
 from tests.src.db.models.factories import OpportunityFactory, UserFactory, WorkflowFactory
-from tests.workflow.state_machine.test_state_machines import BasicState
-from tests.workflow.workflow_test_util import build_process_workflow_event
+from tests.src.workflow.state_machine.test_state_machines import BasicState
+from tests.src.workflow.workflow_test_util import build_process_workflow_event
 
 logger = logging.getLogger(__name__)
